Comtinue refactoring Style into query vs styles
