+++ api/web/src/components/Layer/Styles.vue
@@ -17,12 +17,12 @@
                 <div v-if='global_enabled' class="d-flex justify-content-center">
                     <div class="btn-list">
                         <div class="btn-group" role="group">
-                            <input v-model='type' type="radio" class="btn-check" name="type-toolbar" value='basic'>
-                            <label @click='type="basic"' class="btn btn-icon px-3">
+                            <input v-model='mode' type="radio" class="btn-check" name="type-toolbar" value='basic'>
+                            <label @click='mode="basic"' class="btn btn-icon px-3">
                                 <AbcIcon/> Basic
                             </label>
-                            <input v-model='type' type="radio" class="btn-check" name="type-toolbar" value='query'>
-                            <label @click='type="query"' class="btn btn-icon px-3">
+                            <input v-model='mode' type="radio" class="btn-check" name="type-toolbar" value='query'>
+                            <label @click='mode="query"' class="btn btn-icon px-3">
                                 <CodeIcon/> Query
                             </label>
                         </div>
@@ -30,6 +30,11 @@
                 </div>
             </div>
             <div class='col-md-4'>
+                <div class='d-flex'>
+                    <div class='ms-auto'>
+                        <button v-if='mode === "query"' @click='query = {}' class='btn'>New Query</button>
+                    </div>
+                </div>
             </div>
         </div>
     </div>
@@ -37,126 +42,37 @@
     <div v-if='!global_enabled' class='card-body'>
         Style Overrides are disabled
     </div>
-    <div v-else-if='type === "query" && !query' class='card-body'>
-        <div class="list-group list-group-flush">
-            <div :key='q_idx' v-for='(q, q_idx) in queries'>
-                <a @click='query = q_idx' class="cursor-pointer list-group-item list-group-item-action" v-text='q.query'></a>
+    <template v-else-if='mode === "query" && !query'>
+        <div class='card-body'>
+            <div class="list-group list-group-flush">
+                <div :key='q_idx' v-for='(q, q_idx) in queries'>
+                    <a @click='query = q_idx' class="cursor-pointer list-group-item list-group-item-action" v-text='q.query'></a>
+                </div>
             </div>
         </div>
-    </div>
-    <div v-else class='card-body'>
-        <div class='row'>
-            <div class="d-flex justify-content-center mb-4">
-                <div class="btn-list">
-                    <div class="btn-group" role="group">
-                        <input v-model='mode' type="radio" class="btn-check" name="geom-toolbar" value='point'>
-                        <label @click='mode="point"' class="btn btn-icon px-3">
-                            <PointIcon/> Points
-                        </label>
-                        <input v-model='mode' type="radio" class="btn-check" name="geom-toolbar" value='line'>
-                        <label @click='mode="line"' class="btn btn-icon px-3">
-                            <LineIcon/> Lines
-                        </label>
-                        <input v-model='mode' type="radio" class="btn-check" name="geom-toolbar" value='polygon'>
-                        <label @click='mode="polygon"' class="btn btn-icon px-3">
-                            <PolygonIcon/> Polygons
-                        </label>
-                    </div>
-                </div>
-            </div>
+    </template>
+    <template v-else-if='mode === "query" && typeof query === "object"'>
+        <div class='card-body'>
+            <TablerInput v-model='query.query' placeholder='JSONata Query' label='JSONata Query'/>
 
-            <div v-if='filters[mode].color !== undefined' class='col-md-6 mb-3'>
-                <label class="form-label">Point Color</label>
-                <div class="row g-2">
-                    <div :key='color' v-for='color in [
-                        "dark", "white", "blue", "azure", "indigo", "purple", "pink", "red", "orange", "yellow", "lime"
-                    ]'
-                    class="col-auto">
-                        <label class="form-colorinput">
-                            <input :disabled='disabled' v-model='filters[mode].color' :value='color' type="radio" class="form-colorinput-input">
-                            <span class="form-colorinput-color bg-dark" :class='[
-                                `bg-${color}`
-                            ]'></span>
-                        </label>
-                    </div>
-                </div>
-            </div>
-
-            <div v-if='filters[mode].stroke !== undefined' class='col-md-6 mb-3'>
-                <label class="form-label">Line Color</label>
-                <div class="row g-2">
-                    <div :key='color' v-for='color in [
-                        "dark", "white", "blue", "azure", "indigo", "purple", "pink", "red", "orange", "yellow", "lime"
-                    ]'
-                    class="col-auto">
-                        <label class="form-colorinput">
-                            <input :disabled='disabled' v-model='filters[mode].stroke' :value='color' type="radio" class="form-colorinput-input">
-                            <span class="form-colorinput-color bg-dark" :class='[
-                                `bg-${color}`
-                            ]'></span>
-                        </label>
-                    </div>
-                </div>
-            </div>
-
-            <div v-if='filters[mode]["stroke-style"] !== undefined' class='col-md-6 mb-3'>
-                <label class="form-label">Line Style</label>
-                <select :disabled='disabled' v-model='filters[mode]["stroke-style"]' class="form-select">
-                    <option value="solid">Solid</option>
-                    <option value="dashed">Dashed</option>
-                    <option value="dotted">Dotted</option>
-                    <option value="outlined">Outlined</option>
-                </select>
-            </div>
-            <div v-if='filters[mode]["stroke-width"] !== undefined' class='col-md-6 mb-3'>
-                <label class="form-label">Line Thickness</label>
-                <input :disabled='disabled' v-model='filters[mode]["stroke-width"]' type="range" class="form-range mb-2" min="1" max="6" step="1">
-            </div>
-            <div v-if='filters[mode]["stroke-opacity"] !== undefined' class='col-md-6 mb-3'>
-                <label class="form-label">Line Opacity</label>
-                <input :disabled='disabled' v-model='filters[mode]["stroke-opacity"]' type="range" class="form-range mb-2" min="0" max="256" step="1">
-            </div>
-
-            <div v-if='filters[mode].fill !== undefined' class='col-md-6 mb-3'>
-                <label class="form-label">Fill Color</label>
-                <div class="row g-2">
-                    <div :key='color' v-for='color in [
-                        "dark", "white", "blue", "azure", "indigo", "purple", "pink", "red", "orange", "yellow", "lime"
-                    ]'
-                    class="col-auto">
-                        <label class="form-colorinput">
-                            <input :disabled='disabled' v-model='filters[mode].fill' :value='color' type="radio" class="form-colorinput-input">
-                            <span class="form-colorinput-color bg-dark" :class='[
-                                `bg-${color}`
-                            ]'></span>
-                        </label>
-                    </div>
-                </div>
-            </div>
-            <div v-if='filters[mode]["fill-opacity"] !== undefined' class='col-md-6 mb-3'>
-                <label class="form-label">Fill Opacity</label>
-                <input :disabled='disabled' v-model='filters[mode]["fill-opacity"]' type="range" class="form-range mb-2" min="0" max="256" step="1">
-            </div>
-
-            <div class='col-md-12'>
-                <TablerInput :disabled='disabled' v-model='filters[mode].remarks' label='Remarks'/>
-            </div>
+            <StylesSingle v-model='query'/>
         </div>
-    </div>
+    </template>
+    <template v-else>
+        <StylesSingle v-model='basic'/>
+    </template>
 </div>
 </template>
 
 <script>
 import {
-    TablerInput
-} from '@tak-ps/vue-tabler'
-import {
     AbcIcon,
-    PointIcon,
-    LineIcon,
-    PolygonIcon,
     CodeIcon
 } from 'vue-tabler-icons'
+import {
+    TablerInput
+} from '@tak-ps/vue-tabler';
+import StylesSingle from './Styles/Single.vue';
 
 export default {
     name: 'StyleUtil',
@@ -179,117 +95,30 @@ export default {
     },
     data: function() {
         return {
-            type: 'basic',
-            mode: 'point',
+            mode: 'basic',
             global_enabled: null,
             query: null,
             queries: [],
-            filters: {
-                point: {
-                    color: 'red',
-                    remarks: ''
-                },
-                line: {
-                    stroke: 'red',
-                    'stroke-style': 'solid',
-                    'stroke-opacity': 256,
-                    'stroke-width': 3,
-                    remarks: ''
-                },
-                polygon: {
-                    stroke: 'red',
-                    'stroke-style': 'solid',
-                    'stroke-opacity': 256,
-                    'stroke-width': 3,
-                    'fill': 'red',
-                    'fill-opacity': 256,
-                    remarks: ''
-                }
-            }
+            basic: {}
         };
     },
     watch: {
         global_enabled: function() {
             this.$emit('enabled', this.global_enabled);
-        },
-        filters: {
-            deep: true,
-            handler: function() {
-                this.format();
-            }
         }
     },
     mounted: function() {
         this.global_enabled = this.enabled;
 
-        for (const key in this.modelValue) {
-            const style = JSON.parse(JSON.stringify(this.modelValue[key]));
+        if (this.modelValue.queries) this.queries = this.modelValue.queries;
+        else this.basic = this.modelValue;
 
-            const colors = {
-                '#1d273b': 'dark',
-                '#ffffff': 'white',
-                '#206bc4': 'blue',
-                '#4299e1': 'azure',
-                '#4263eb': 'indigo',
-                '#ae3ec9': 'purple',
-                '#d6336c': 'pink',
-                '#d63939': 'red',
-                '#f76707': 'orange',
-                '#f59f00': 'yellow',
-                '#74b816': 'lime'
-            };
-
-            for (const color of ['color', 'stroke', 'fill']) {
-                if (style[color]) style[color] = colors[style[color]];
-            }
-
-            Object.assign(this.filters[key], style);
-
-            this.format();
-        }
-    },
-    methods: {
-        format: function() {
-            const styles = JSON.parse(JSON.stringify(this.filters));
-            if (styles) {
-                const colors = {
-                    dark: '#1d273b',
-                    white: '#ffffff',
-                    blue: '#206bc4',
-                    azure: '#4299e1',
-                    indigo: '#4263eb',
-                    purple: '#ae3ec9',
-                    pink: '#d6336c',
-                    red: '#d63939',
-                    orange: '#f76707',
-                    yellow: '#f59f00',
-                    lime: '#74b816'
-                };
-
-                for (const key in styles) {
-                    for (const intkey of ['fill-opacity', 'stroke-width', 'stroke-opacity']) {
-                        if (styles[key][intkey]) styles[key][intkey] = parseInt(styles[key][intkey])
-                    }
-
-                    for (const color of ['color', 'stroke', 'fill']) {
-                        if (styles[key][color]) {
-                            styles[key][color] = colors[styles[key][color]];
-                        }
-                    }
-                }
-            }
-
-
-            this.$emit('update:modelValue', styles);
-        }
     },
     components: {
-        TablerInput,
-        PointIcon,
-        LineIcon,
-        PolygonIcon,
         CodeIcon,
         AbcIcon,
+        StylesSingle,
+        TablerInput
     }
 }
 </script>
